kafka/consumer/requirements.txt
kafka-python==2.0.2
python-decouple==3.6
